hw/block/nvme: A namespace's I/O command set is fixed
The only thing that determines a namespace's I/O command set is the CSI.

It should not be set to different things depending on CC.CSS register.
Instead, nvme_io_cmd() and nvme_admin_cmd() should have the proper
checking of CC.CSS register.

Likewise, if a namespace is attached or not, does not affect the IOCS
of the namespace. The checking should be done in nvme_io_cmd() and
nvme_admin_cmd(). Admin commands that are always allowed, regardless of
CC.CSS (e.g. nvme_identify_ns()), already have logic to handle detached
namespaces.

(Commit has not been compile tested)

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>