hw/ide: Add General Purpose Logging (GPL) feature set support
Since the log pages that READ LOG DMA EXT will read will only be
in RAM, and not on a BlockBackend, the READ LOG DMA EXT command
will simply use the same implementation as READ LOG EXT, which
does a simple memcpy().

While this is not strictly correct, this significantly simplifies
things, and since the data we return will be the same, and since
the host driver will not be able to notice how we transferred the
data, there should be no downside of implementing it this way.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>